igen igen.....

+++ exercise1.c
@@ -20,7 +20,7 @@
  * Returns: approximation of sin(x)
  */
 double factorial(int n) {
-    int f = 1;
+    double f = 1;
     for (int i = 2; i <= n; i++) {
         f *= i;
     }
